refactor: take "Meta" out of command class names
