+++ src/Commands/Sync/DumpPluginsCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Commands\Sync\Meta;
+namespace App\Commands\Sync;
 
 use App\Commands\AbstractBaseCommand;
 use App\Services\Metadata\PluginMetadataService;
@@ -11,7 +11,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class MetaDumpPluginsCommand extends AbstractBaseCommand
+class DumpPluginsCommand extends AbstractBaseCommand
 {
     public function __construct(
         private PluginMetadataService $meta,
