+++ src/Commands/Sync/FetchPluginsCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Commands\Sync\Meta;
+namespace App\Commands\Sync;
 
 use App\Integrations\Wordpress\PluginRequest;
 use App\Integrations\Wordpress\WordpressApiConnector;
@@ -10,7 +10,7 @@ use App\ResourceType;
 use App\Services\Metadata\PluginMetadataService;
 use Saloon\Http\Request;
 
-class MetaFetchPluginsCommand extends AbstractMetaFetchCommand
+class FetchPluginsCommand extends AbstractFetchCommand
 {
     public function __construct(
         PluginMetadataService $meta,
